moved image code out
